+++ src/boomi/services/async_/shared_communication_channel_component.py
@@ -5,6 +5,7 @@ from ..shared_communication_channel_component import (
     SharedCommunicationChannelComponentService,
 )
 from ...models import (
+    SharedCommunicationChannelComponent,
     SharedCommunicationChannelComponentBulkRequest,
     SharedCommunicationChannelComponentQueryResponse,
     SharedCommunicationChannelComponentQueryConfig,
@@ -36,6 +37,25 @@ class SharedCommunicationChannelComponentServiceAsync(
         return to_async(super().update_shared_communication_channel_component)(
             id_, request_body
         )
+
+    def create_shared_communication_channel_component_json(
+        self, request_body: Union[SharedCommunicationChannelComponent, dict] = None
+    ) -> Awaitable[Union[SharedCommunicationChannelComponent, str, dict]]:
+        return to_async(
+            super().create_shared_communication_channel_component_json
+        )(request_body)
+
+    def get_shared_communication_channel_component_json(
+        self, id_: str
+    ) -> Awaitable[Union[SharedCommunicationChannelComponent, str, dict]]:
+        return to_async(super().get_shared_communication_channel_component_json)(id_)
+
+    def update_shared_communication_channel_component_json(
+        self, id_: str, request_body: Union[SharedCommunicationChannelComponent, dict] = None
+    ) -> Awaitable[Union[SharedCommunicationChannelComponent, str, dict]]:
+        return to_async(
+            super().update_shared_communication_channel_component_json
+        )(id_, request_body)
 
     def delete_shared_communication_channel_component(
         self, id_: str